requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
gunicorn==20.1.0
idna==3.2
PyJWT==1.7.1
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.6
